backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
jaclang>=0.5.0
setuptools>=65.0.0
litellm>=1.17.0
python-dotenv==1.0.0
google-generativeai>=0.3.0
protobuf==4.25.5
